requirements.txt
Flask==3.0.0
Pillow==10.1.0
numpy==1.26.2
matplotlib==3.8.2
Werkzeug==3.0.1
gunicorn==21.2.0
